refactor: make main model generic
move queueing screen to its own model and create a generic model in main which can be used for any model as more are added
